+++ core/src/inline_layout/line_metrics.hpp
@@ -9,8 +9,8 @@ struct LineMetrics {
     float baseline;
     float x_height;
 
-    LineMetrics add(int added);
-    LineMetrics scale(int ratio);
+    LineMetrics add(float added);
+    LineMetrics scale(float ratio);
 
     static LineMetrics from_paint(const SkPaint& paint);
 };
